+++ shared/src/commonMain/kotlin/ui/composables/DCQLCredentialQuerySubmissionSelectionOption.kt
@@ -25,7 +25,6 @@ import at.asitplus.wallet.app.common.thirdParty.kotlinx.serialization.json.leafN
 import at.asitplus.wallet.lib.agent.SdJwtValidator
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.data.CredentialToJsonConverter.toJsonElement
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.jws.SdJwtSigned
 import data.Attribute
@@ -44,18 +43,16 @@ fun DCQLCredentialQuerySubmissionSelectionOption(
     isSelected: Boolean,
     onToggleSelection: () -> Unit,
     option: DCQLCredentialSubmissionOption<SubjectCredentialStore.StoreEntry>,
-    checkRevocationStatus: suspend (SubjectCredentialStore.StoreEntry) -> TokenStatus?,
+    checkCredentialFreshness: suspend (SubjectCredentialStore.StoreEntry) -> CredentialFreshnessSummary,
     decodeToBitmap: (ByteArray) -> ImageBitmap?,
     modifier: Modifier = Modifier,
 ) {
-    val credentialStatusState by produceState(
-        CredentialStatusState.Loading as CredentialStatusState,
+    val credentialFreshnessValidationState by produceState(
+        CredentialFreshnessValidationState.Loading as CredentialFreshnessValidationState,
         option.credential
     ) {
-        value = CredentialStatusState.Loading
-        value = CredentialStatusState.Success(
-            checkRevocationStatus(option.credential)
-        )
+        value = CredentialFreshnessValidationState.Loading
+        value = CredentialFreshnessValidationState.Done(checkCredentialFreshness(option.credential))
     }
 
 
@@ -106,13 +103,13 @@ fun DCQLCredentialQuerySubmissionSelectionOption(
     }
 
     CredentialSelectionCardLayout(
-        credentialStatusState = credentialStatusState,
+        credentialFreshnessValidationState = credentialFreshnessValidationState,
         onClick = onToggleSelection,
         isSelected = isSelected,
         modifier = modifier,
     ) {
         CredentialSelectionCardHeader(
-            credentialStatusState = credentialStatusState,
+            credentialFreshnessValidationState = credentialFreshnessValidationState,
             credential = credential,
             modifier = Modifier.fillMaxWidth()
         )