+++ shared/src/commonMain/kotlin/ui/composables/TimelinessIndicator.kt
@@ -1,0 +1,8 @@
+package ui.composables
+
+interface TimelinessIndicator {
+    val isExpired: Boolean
+    val isNotYetValid: Boolean
+    val isTimely: Boolean
+        get() = !isNotYetValid && !isExpired
+}