+++ shared/src/commonMain/kotlin/ui/composables/credentials/CredentialSelectionCardHeader.kt
@@ -4,16 +4,11 @@ import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import at.asitplus.valera.resources.Res
-import at.asitplus.valera.resources.error_credential_status_invalid
 import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.agent.representation
 import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.iconLabel
 import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.uiLabel
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
-import org.jetbrains.compose.resources.stringResource
-import ui.composables.BigErrorText
-import ui.composables.CredentialStatusState
+import ui.composables.CredentialFreshnessValidationState
 import ui.composables.LabeledText
 import ui.composables.PersonAttributeDetailCardHeading
 import ui.composables.PersonAttributeDetailCardHeadingIcon
@@ -21,7 +16,7 @@ import ui.composables.PersonAttributeDetailCardHeadingIcon
 
 @Composable
 fun ColumnScope.CredentialSelectionCardHeader(
-    credentialStatusState: CredentialStatusState,
+    credentialFreshnessValidationState: CredentialFreshnessValidationState,
     credential: SubjectCredentialStore.StoreEntry,
     modifier: Modifier = Modifier,
 ) {
@@ -36,12 +31,11 @@ fun ColumnScope.CredentialSelectionCardHeader(
             )
         },
         actionButtons = {
-            when (credentialStatusState) {
-                CredentialStatusState.Loading -> CircularProgressIndicator()
-                is CredentialStatusState.Success -> when (credentialStatusState.tokenStatus) {
-                    TokenStatus.Invalid -> BigErrorText(stringResource(Res.string.error_credential_status_invalid))
-                }
+            when(val it = credentialFreshnessValidationState) {
+                CredentialFreshnessValidationState.Loading -> CircularProgressIndicator()
+                is CredentialFreshnessValidationState.Done -> MainCredentialIssue(it.credentialFreshnessSummary)
             }
         }
     )
 }
+